requirements.txt
openai==1.12.0
python-dotenv==1.0.0
rich==13.7.0
typer==0.9.0
pyyaml==6.0.1
pandas==2.2.0
textual==0.38.1
asciimatics==1.14.0
questionary==2.0.1
alive-progress==3.1.5
art==6.0
termcolor==2.3.0 